requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.3
asgiref==3.5.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.5.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-crontab==0.7.1
django-environ==0.8.1
django-filter==21.1
django-import-export==2.8.0
django-livereload-server==0.4
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
djlint==0.7.5
dynaconf==3.1.7
environ==1.0
et-xmlfile==1.1.0
filelock==3.6.0
future==0.18.2
html5lib==1.1
httpie==3.1.0
hvac==0.11.2
idna==3.3
importlib-metadata==4.11.2
install==1.3.5
jwt==1.3.1
knox==0.1.14
kombu==5.2.4
loguru==0.6.0
lxml==4.8.0
Markdown==3.3.6
MarkupPy==1.14
multidict==6.0.2
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.2
passlib==1.7.4
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
pyotp==2.6.0
PyPDF3==1.0.6
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.3.2
reportlab==3.6.8
requests==2.27.1
requests-toolbelt==0.9.1
shortuuid==1.0.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.1
tinycss2==1.1.1
tomlkit==0.9.2
tornado==6.1
tqdm==4.63.0
tzdata==2022.1
urllib3==1.26.8
validators==0.18.2
vine==5.0.0
virtualenv==20.13.3
wcwidth==0.2.5
webencodings==0.5.1
win32-setctime==1.1.0
xhtml2pdf==0.2.6
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
